Final preps
General preps for the final version of 3.1.3 which includes fixes and
new additions such as more hospitals, EMS vehicle & Item shops and other
things.

Note that the EMS Heli shop spawns the helicopter on the helipad of the
hospital and due to complications that I don't want to be bothered with
the helicopters can only be 'rented' and not bought for storage.