3.1.3 RC1
A release candidate for 3.1.3. This includes the medic siren lights and
support for siren sound (not included but the code is there) as well as
adding the ATM's to the hospitals.

The Airport lamps at the Kavala Cop HQ have had their simulation
re-enabled so that the lights work at night.