More roundup of fixes.
Fixed issues with last few commits and added a new feature that I am
debating with which is security camera like functionality. Currently you
can access a camera feed from the Kavala police HQ from a laptop.

Time Module was changed up so that only the server skips the time
forward and not the client.

Fixed: Weapons dropped on the ground are deleted when a unit dies.
Changed: Medics can now 'Buy' helicopters and pull them from the garage.

Debug message for the Async handler added, will outprint to the RPT the
query and how long it took to execute the query.

Money glitching out when being picked up rapidly with other items also
fixed.