Changes from another day
Fixed: When a cop would repair the vault they could not use the player
menu / interact with anything.
Fixed: inUse variable is unset for various exist conditions for picking
up money (This stopped you from picking up money relating to a 3.1.3
change or 3.1.2 can't remember).

Removed junk variables and functions that are no longer in use.

Added debug logs for queries that have to wait for a returning value (If
you are suffering from long query load times please report it with the
rpt log!).