4.0 Framework changes
This is most likely the last thing I ever give to the public ever again.
This contains a change over in the framework design and core
configuration. Macros are used A LOT in this build and that is to get
the lazy fucking people away from using my work. If you want to modify
it then guess what.. You're putting the time into learning a little bit
more.

The code changes are no where close to complete and there are problems
still lingering in this build..

Spyglass variable checker has also been changed over to scan ALL
variables set in all namespaces and the variable list is not complete
but I would assume Yuka will be correcting that later on.

Nothing of what I really set out to do for this build is completed or
even done and I have ripped out features that were meant to be added.