Life: Changes for Tanoa compatibility! SEE DESCRIPTION BEFORE MERGING!
Tanoa map is going to be released in the next few months. The idea is to
port Life mod to it.
So, these changes are just for compatibility with Tanoa.

Custom ini will be compatible with both, that's why the name armalife.

Database will need to be different, so, now it has the map in its name.

Also, the MySQL field in EXTDB2 has a new name.
Now it's ArmaLife. This happened because life_server will be compatible
with both maps.