Mission: Port this shit to EDEN format (Another shit...)
Now it should work with EDEN.
Probably there'll be some bugs. Like, vehicles not spawning, player
respawn and NPCs placed into the ground.
BUT, this should work, and it took me some time to do.

Also, if you're going to port it to your server, you should open the
mission with EDEN and save it again as binarized. It'll make the loading
process faster, but, you won't be able to edit the mission using the
text editor anymore. (You can revert it by saving again without the
binary checkbox selected)