Request Medic Fix (#803)
Fixes the issues with having to click request medic twice to get it to
work. Basically if you clicked it in the first 5 seconds it wouldn't
work. Now it waits the respawnDelay specified in the mission.sqm and
will now only enable when it's actually going to work. Haha. Credits to
BoGuu, I sort of helped...