Update functions.sqf
Quickly went through the code and made some adjustments. I tested this locally on my own, worked fine for my own marker. Needs to be tested multiplayer + vehicles etc. Also little concerned of how a crew of people would look as i left the code as it was but it looks a little funny to me.. also. any reasons why we are using (for ""_i"" from 0 to 1 step 0 do ) instead of like a while statement?