server: vars: clean it
extDB error was never being shown, since "life_server_isReady" wouldn't
be set true if an error happened:

https://github.com/AsYetUntitled/Framework/blob/33213795c08f082fecd048ced786f06e3580f3c4/life_server/init.sqf#L62

This way, it'll be shown to client if extDB failed to establish a
connection. Also, it's better to set public a var that never changes its
kind (Before, it could be a string or an array).