init: fix "waiting for the server to be ready" bug. (#217)
* init: fix "waiting for the server to be ready" bug.

Waiting for a var that'll never be set again isn't smart.
This fixes a possible hang in the AL init for the first player to join the
mission. (If the server isn't configured to auto initialize)

* HC: Vars: Clean it

If the removed vars were set, then, life_HC_isActive was true.

The "core\init.sqf" checks were being done by HC:

https://github.com/AsYetUntitled/Framework/blob/94b0e526180e7c89f0f4974544600ccf30eaf433/life_hc/initHC.sqf#L52

* server: vars: clean it

extDB error was never being shown, since "life_server_isReady" wouldn't
be set true if an error happened:

https://github.com/AsYetUntitled/Framework/blob/33213795c08f082fecd048ced786f06e3580f3c4/life_server/init.sqf#L62

This way, it'll be shown to client if extDB failed to establish a
connection. Also, it's better to set public a var that never changes its
kind (Before, it could be a string or an array).

* spyglass: case

* init: optimize

* lazy eval